+++ docs/conf.py
@@ -31,7 +31,7 @@ MOCK_MODULES = [
     "requests",
     "bs4",
     "dota2py",
-    "novaclient.v2",
+    "novaclient",
     "speedtest_cli",
     "pyzabbix",
     "vk",
@@ -39,8 +39,11 @@ MOCK_MODULES = [
     "dateutil",
     "httplib2",
     "oauth2client",
-    "apiclient"
-
+    "apiclient",
+    "googleapiclient.errors",
+    "vlc",
+    "dateutil.tz",
+    "i3ipc",
 ]
 
 for mod_name in MOCK_MODULES: