+++ i3pystatus/github.py
@@ -1,14 +1,24 @@
-from i3pystatus import IntervalModule
-import requests
 import json
+
+import requests
+from i3pystatus import IntervalModule
+from i3pystatus import logger
 from i3pystatus.core import ConfigError
 from i3pystatus.core.util import user_open, internet, require
+from requests import Timeout, ConnectionError
 
 
 class Github(IntervalModule):
     """
-    Check Github for pending notifications.
+    Check GitHub for pending notifications.
     Requires `requests`
+
+    Availables authentication methods:
+
+    * username + password
+    * access_token (manually generate a new token at https://github.com/settings/tokens)
+
+    See https://developer.github.com/v3/#authentication for more informations.
 
     Formatters:
 
@@ -22,6 +32,7 @@ class Github(IntervalModule):
     color = '#78EAF2'
     username = ''
     password = ''
+    access_token = ''
     format = '{unread}'
     interval = 600
     keyring_backend = None
@@ -34,6 +45,7 @@ class Github(IntervalModule):
         ('unread_marker', 'sets the string that the "unread" formatter shows when there are pending notifications'),
         ("username", ""),
         ("password", ""),
+        ("access_token", "see https://developer.github.com/v3/#authentication"),
         ("color", "")
     )
 
@@ -42,16 +54,23 @@ class Github(IntervalModule):
 
     @require(internet)
     def run(self):
-        format_values = dict(unread_count='', unread='')
 
-        response = requests.get('https://api.github.com/notifications', auth=(self.username, self.password))
-        data = json.loads(response.text)
+        try:
+            if self.access_token:
+                response = requests.get('https://api.github.com/notifications?access_token=' + self.access_token)
+            else:
+                response = requests.get('https://api.github.com/notifications', auth=(self.username, self.password))
+            data = json.loads(response.text)
+        except (ConnectionError, Timeout) as e:
+            logger.warn(e)
+            data = []
 
         # Bad credentials
         if isinstance(data, dict):
             err_msg = data['message']
             raise ConfigError(err_msg)
 
+        format_values = dict(unread_count='', unread='')
         unread = len(data)
         if unread > 0:
             format_values['unread_count'] = unread