+++ i3pystatus/updates/pacman.py
@@ -12,6 +12,12 @@ class Pacman(Backend):
     def updates(self):
         command = ["checkupdates"]
         checkupdates = run_through_shell(command)
-        return checkupdates.out.count('\n')
+        return checkupdates.out.count("\n"), checkupdates.out
 
 Backend = Pacman
+
+if __name__ == "__main__":
+    """
+    Call this module directly; Print the update count and notification body.
+    """
+    print("Updates: {}\n\n{}".format(*Backend().updates))